feat(postgres): implement real tree extensions on db

+++ src/Grimoire.Application/Dto/Book/ChapterListResponseDto.cs
@@ -2,7 +2,7 @@ namespace Grimoire.Application.Dto.Book;
 
 public class ChapterListResponseDto {
 	public string VolumeId { get; init; } = string.Empty;
-	public float Order { get; init; }
+	public double Order { get; init; }
 	public string Title { get; init; } = string.Empty;
 	public string Id { get; init; } = string.Empty;
 	public DateTime? UpdatedAt { get; init; }
