+++ src/Grimoire.Application/Dto/Book/Tree/BookTreeDto.cs
@@ -13,7 +13,7 @@ public sealed record BookTreeNodeDto {
 	public required string Id { get; init; }
 	public required BookTreeNodeType Type { get; init; }
 	public required string Title { get; init; }
-	public float? Order { get; init; }
+	public double? Order { get; init; }
 	public string? ParentId { get; init; }
 	public List<BookTreeNodeDto> Children { get; init; } = [];
 }
