+++ src/Grimoire.Application/Dto/Book/UpdateVolumeRequestDto.cs
@@ -2,4 +2,4 @@ namespace Grimoire.Application.Dto.Book;
 
 using Metadata;
 
-public record UpdateVolumeRequestDto(float? Order, string? Title, VolumeMetadataDto? Metadata, string? SeriesId = null);
+public record UpdateVolumeRequestDto(double? Order, string? Title, VolumeMetadataDto? Metadata, string? SeriesId = null);
