+++ src/Grimoire.Application/Service/Contract/IBookTreeService.cs
@@ -27,15 +27,16 @@ public interface IChapterNodeService {
 }
 
 public interface INodeManagerService {
-	public Task<BookNodeModel> CreateNode(Guid id, BookNodeType type, Guid? parentId, string title, float order, CancellationToken cancellationToken = default);
-	public Task<BookNodeModel> UpdateNode(Guid id, string? title, float? order, CancellationToken cancellationToken = default);
-	public Task MoveNode(Guid nodeId, Guid? newParentId, float newOrder, CancellationToken cancellationToken = default);
+	public Task<BookNodeModel> CreateNode(Guid id, BookNodeType type, Guid? parentId, string title, double order, CancellationToken cancellationToken = default);
+	public Task<BookNodeModel> UpdateNode(Guid id, string? title, double? order, CancellationToken cancellationToken = default);
+	public Task MoveNode(Guid nodeId, Guid? newParentId, double newOrder, CancellationToken cancellationToken = default);
 	public Task<int> DeleteSubtree(Guid nodeId, CancellationToken cancellationToken = default);
 	public Task ReconcileNodesBulk(
 		Guid seriesId,
-		List<(Guid Id, BookNodeType Type, Guid? ParentId, string Title, float Order)> nodes,
+		List<(Guid Id, BookNodeType Type, Guid? ParentId, string Title, double Order)> nodes,
 		CancellationToken cancellationToken = default);
 }
+
 
 public interface IBookTreeService : ISeriesNodeService, IVolumeNodeService, IChapterNodeService, INodeManagerService {
 }
